Moved Load* methods on IQueryExecutor to extension methods, to make it easier for users to provide their own implementation of IQueryExecutor
