+++ source/Nevermore/IQueryExecutor.cs
@@ -83,49 +83,7 @@ namespace Nevermore
         /// <returns>A builder to further customize the query</returns>
         IDeleteQueryBuilder<TDocument> DeleteQuery<TDocument>() where TDocument : class, IId;
 
-        /// <summary>
-        /// Loads a single document given its ID. If the item is not found, returns <c>null</c>.
-        /// </summary>
-        /// <typeparam name="TDocument">The type of document being queried. Results from the database will be mapped to this type.</typeparam>
-        /// <param name="id">The <c>Id</c> of the document to find.</param>
-        /// <returns>The document, or <c>null</c> if the document is not found.</returns>
-        TDocument Load<TDocument>(string id) where TDocument : class, IId;
-
-        /// <summary>
-        /// Loads a set of documents by their ID's. Documents that are not found are excluded from the result list (that is,
-        /// the results may contain less items than the number of ID's queried for).
-        /// </summary>
-        /// <typeparam name="TDocument">The type of document being queried. Results from the database will be mapped to this type.</typeparam>
-        /// <param name="ids">A collection of ID's to query by.</param>
-        /// <returns>The documents.</returns>
-        TDocument[] Load<TDocument>(IEnumerable<string> ids) where TDocument : class, IId;
-
-        /// <summary>
-        /// Loads a set of documents by their ID's. Documents that are not found are excluded from the result list (that is,
-        /// the results may contain less items than the number of ID's queried for).
-        /// </summary>
-        /// <typeparam name="TDocument">The type of document being queried. Results from the database will be mapped to this type.</typeparam>
-        /// <param name="ids">A collection of ID's to query by.</param>
-        /// <returns>The documents as a lazy loaded stream.</returns>
-        IEnumerable<TDocument> LoadStream<TDocument>(IEnumerable<string> ids) where TDocument : class, IId;
-
-        /// <summary>
-        /// Loads a single document given its ID. If the item is not found, throws a <see cref="ResourceNotFoundException" />.
-        /// </summary>
-        /// <typeparam name="TDocument">The type of document being queried. Results from the database will be mapped to this type.</typeparam>
-        /// <param name="id">The <c>Id</c> of the document to find.</param>
-        /// <returns>The document, or <c>null</c> if the document is not found.</returns>
-        TDocument LoadRequired<TDocument>(string id) where TDocument : class, IId;
-
-        /// <summary>
-        /// Loads a set of documents by their ID's. If any of the documents are not found, a
-        /// <see cref="ResourceNotFoundException" /> will be thrown.
-        /// </summary>
-        /// <typeparam name="TDocument">The type of document being queried. Results from the database will be mapped to this type.</typeparam>
-        /// <param name="ids">A collection of ID's to query by.</param>
-        /// <returns>The documents.</returns>
-        TDocument[] LoadRequired<TDocument>(IEnumerable<string> ids) where TDocument : class, IId;
-
+        
         /// <summary>
         /// Immediately inserts a new item into the default table for the document type. The item will have an automatically
         /// assigned ID, and that ID value will be visible in the <code>Id</code> property of the object as soon as
